Moved MSALApiDispatcher to internal package
