+++ msal/src/androidTest/java/com/microsoft/identity/client/PublicClientApplicationTest.java
@@ -38,6 +38,7 @@ import android.support.test.runner.AndroidJUnit4;
 import android.test.AndroidTestCase;
 
 import com.microsoft.identity.client.internal.MsalUtils;
+import com.microsoft.identity.client.internal.controllers.MSALApiDispatcher;
 import com.microsoft.identity.client.internal.controllers.RequestCodes;
 import com.microsoft.identity.common.internal.net.HttpUrlConnectionFactory;
 
@@ -733,7 +734,7 @@ public final class PublicClientApplicationTest extends AndroidTestCase {
 
             final Intent resultIntent = new Intent();
             resultIntent.putExtra(Constants.AUTHORIZATION_FINAL_URL, getFinalAuthUrl());
-            com.microsoft.identity.client.MSALApiDispatcher.completeInteractive(RequestCodes.LOCAL_AUTHORIZATION_REQUEST,
+            MSALApiDispatcher.completeInteractive(RequestCodes.LOCAL_AUTHORIZATION_REQUEST,
                     Constants.UIResponse.AUTH_CODE_COMPLETE, resultIntent);
 
             resultLock.await();
