+++ msal/src/main/java/com/microsoft/identity/client/internal/controllers/MSALApiDispatcher.java
@@ -20,11 +20,14 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.client;
+package com.microsoft.identity.client.internal.controllers;
 
 import android.content.Intent;
 import android.os.Handler;
 
+import com.microsoft.identity.client.AuthenticationResult;
+import com.microsoft.identity.client.MsalException;
+import com.microsoft.identity.client.MsalUserCancelException;
 import com.microsoft.identity.client.internal.controllers.AcquireTokenResult;
 import com.microsoft.identity.client.internal.controllers.ExceptionAdapter;
 import com.microsoft.identity.client.internal.controllers.MSALInteractiveTokenCommand;
